# `remove` vectorization