--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -4515,6 +4515,30 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(sentinel_for<_Se, _It>);
             _STL_INTERNAL_STATIC_ASSERT(indirect_binary_predicate<ranges::equal_to, projected<_It, _Pj>, const _Ty*>);
 
+#if _USE_STD_VECTOR_ALGORITHMS
+#if _HAS_CXX20
+            if constexpr (_Vector_alg_in_find_is_safe<_It, _Ty> && sized_sentinel_for<_Se, _It>
+                          && is_same_v<_Pj, identity>) {
+                if (!_STD is_constant_evaluated())
+#endif // _HAS_CXX20
+                {
+                    const auto _Size = _Last - _First;
+                    if (!_STD _Could_compare_equal_to_value_type<_It>(_Val)) {
+                        return {_First + _Size, _First + _Size};
+                    }
+                    const auto _First_ptr = _To_address(_First);
+                    const auto _Last_ptr  = _First_ptr + static_cast<size_t>(_Size);
+                    const auto _Result    = _STD _Remove_vectorized(_First_ptr, _Last_ptr, _Val);
+
+                    if constexpr (is_pointer_v<_It>) {
+                        return {static_cast<_It>(_Result), static_cast<_It>(_Last_ptr)};
+                    } else {
+                        return {_First + (_Result - _First_ptr), _First + _Size};
+                    }
+                }
+            }
+#endif // _USE_STD_VECTOR_ALGORITHMS
+
             _First     = _RANGES _Find_unchecked(_STD move(_First), _Last, _Val, _Proj);
             auto _Next = _First;
             if (_First == _Last) {