--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -4589,14 +4589,17 @@ namespace ranges {
             _STL_INTERNAL_STATIC_ASSERT(sentinel_for<_Se, _It>);
             _STL_INTERNAL_STATIC_ASSERT(indirect_binary_predicate<ranges::equal_to, projected<_It, _Pj>, const _Ty*>);
 
+            _First     = _RANGES _Find_unchecked(_STD move(_First), _Last, _Val, _Proj);
+            auto _Next = _First;
+            if (_First == _Last) {
+                return {_STD move(_Next), _STD move(_First)};
+            }
+
 #if _USE_STD_VECTOR_ALGORITHMS
             if constexpr (_Vector_alg_in_find_is_safe<_It, _Ty> && sized_sentinel_for<_Se, _It>
                           && is_same_v<_Pj, identity>) {
                 if (!_STD is_constant_evaluated()) {
-                    const auto _Size = _Last - _First;
-                    if (!_STD _Could_compare_equal_to_value_type<_It>(_Val)) {
-                        return {_First + _Size, _First + _Size};
-                    }
+                    const auto _Size      = _Last - _First;
                     const auto _First_ptr = _STD _To_address(_First);
                     const auto _Last_ptr  = _First_ptr + static_cast<size_t>(_Size);
                     const auto _Result    = _STD _Remove_vectorized(_First_ptr, _Last_ptr, _Val);
@@ -4609,12 +4612,6 @@ namespace ranges {
                 }
             }
 #endif // _USE_STD_VECTOR_ALGORITHMS
-
-            _First     = _RANGES _Find_unchecked(_STD move(_First), _Last, _Val, _Proj);
-            auto _Next = _First;
-            if (_First == _Last) {
-                return {_STD move(_Next), _STD move(_First)};
-            }
 
             while (++_First != _Last) {
                 if (_STD invoke(_Proj, *_First) != _Val) {