--- a/stl/inc/algorithm
+++ b/stl/inc/algorithm
@@ -4600,7 +4600,7 @@ namespace ranges {
                           && is_same_v<_Pj, identity>) {
                 if (!_STD is_constant_evaluated()) {
                     const auto _Size      = _Last - _First;
-                    const auto _First_ptr = _STD _To_address(_First);
+                    const auto _First_ptr = _STD to_address(_First);
                     const auto _Last_ptr  = _First_ptr + static_cast<size_t>(_Size);
                     const auto _Result    = _STD _Remove_vectorized(_First_ptr, _Last_ptr, _Val);