--- a/stl/inc/xmemory
+++ b/stl/inc/xmemory
@@ -2237,10 +2237,6 @@ _NODISCARD_REMOVE_ALG _CONSTEXPR20 _FwdIt remove(_FwdIt _First, const _FwdIt _La
             if (!_STD is_constant_evaluated())
 #endif // _HAS_CXX20
             {
-                if (!_STD _Could_compare_equal_to_value_type<decltype(_UFirst)>(_Val)) {
-                    return _Last;
-                }
-
                 const auto _First_ptr = _STD _To_address(_UFirst);
                 const auto _Result    = _STD _Remove_vectorized(_First_ptr, _STD _To_address(_ULast), _Val);