--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3808,7 +3808,7 @@ void* __stdcall __std_remove_8(void* _First, void* const _Last, const uint64_t _
         _Advance_bytes(_Stop, _Size_bytes & ~size_t{0x1F});
         do {
             const __m256i _Src            = _mm256_loadu_si256(reinterpret_cast<const __m256i*>(_First));
-            const __m256i _Mask           = _mm256_cmpeq_epi32(_Src, _Match);
+            const __m256i _Mask           = _mm256_cmpeq_epi64(_Src, _Match);
             const unsigned _Bingo_swapped = _mm256_movemask_epi8(_mm256_shuffle_epi8(_Mask, _Dense_shuf));
             const unsigned _Bingo         = (_Bingo_swapped | (_Bingo_swapped >> 28)) & 0xF;
             const __m256i _Shuf           = _mm256_cvtepi8_epi32(_mm_loadu_si64(_Remove_patterns_8._Data[_Bingo]));