… benchmarks/CMakeLists.txt
… benchmarks/inc/lorem.hpp
… benchmarks/src/remove.cpp
… benchmarks/src/replace.cpp
… benchmarks/src/search.cpp
… benchmarks/src/sv_equal.cpp
… tests/std/tests/VSO_0000000_vector_algorithms/test.cpp