@CaseyCarter 2024-09-28
Irony: a PR that adds vectorization entitled "`remove` vectorization".