@AlexGuteniev 2024-09-28
> * **Correctness**. [[algorithms.requirements]/3](https://eel.is/c++draft/algorithms.requirements#3) states: _For purposes of determining the existence of data races, algorithms shall not modify objects referenced through an iterator argument unless the specification requires such modification_. Whereas [[alg.remove]](https://eel.is/c++draft/alg.remove) is vague on how the algorithm should work, I think we should only write to elements that has to be written to

There's a caveat: only the elements before the first removed one are not written. Elements after the resulting range (up to 8 of them) may be overwritten with values of some other elements. That's _valid but unspecified state_ as in [alg.remove]/7

If that's not fine, have to use AVX2 masks, which would be slower, and resort to having only 32-bit and 64-bit versions.