@AlexGuteniev 2024-10-04
Modern AMD data would be interesting. 
To avoid tricky swizzling, I added mixing integers and floats in 39974d1. The overall results are better for me, but this mixing seems to have more penalty on AMD than on Intel