@AlexGuteniev 2024-10-06
> `remove_copy` if it succeeds going to be very similar, but would always use AVX2 mask, and so available only for 32 and 64 bit elements. I can try this within the same PR, although it seems big already.

I'm now seeing multiple solutions how to do `remove_copy` for 8 and 16 bit elements