@StephanTLavavej 2024-10-23
Thanks for the detailed writeup! I have no correctness concerns here - the Standard has a note that the garbage values are valid-but-unspecified, so we're fully within our rights to leave totally random values in there, even if the range originally contained (for example) only `10` and `20`.

"If you want `partition`, you know where to find it."