@StephanTLavavej 2024-10-23
Thanks! :heart_eyes_cat: Pushed changes as usual, good results on my 5950X:

Benchmark                            |  Before |   After | Speedup |
-------------------------------------|--------:|--------:|--------:|
`r<alg_type::std_fn, std::uint8_t>`  | 1291 ns |  360 ns | 3.59
`r<alg_type::std_fn, std::uint16_t>` | 1291 ns |  338 ns | 3.82
`r<alg_type::std_fn, std::uint32_t>` | 1285 ns |  491 ns | 2.62
`r<alg_type::std_fn, std::uint64_t>` | 1504 ns | 1151 ns | 1.31
`r<alg_type::rng, std::uint8_t>`     | 1317 ns |  355 ns | 3.71
`r<alg_type::rng, std::uint16_t>`    | 1250 ns |  330 ns | 3.79
`r<alg_type::rng, std::uint32_t>`    | 1330 ns |  489 ns | 2.72
`r<alg_type::rng, std::uint64_t>`    | 2090 ns | 1082 ns | 1.93