@StephanTLavavej 2024-10-24
I've pushed a merge with `main` to resolve merge conflicts in `search.cpp` (structured as one commit to replicate the stupid thing I did in MSVC, followed by a proper fix, so I can mirror the latter).

Now, this directly constructs `std::string` from the `std::string_view`s `src_haystack` and `src_needle`, and constructs `std::vector` from their `.begin()` and `.end()`. These are stylistic improvements to reduce verbosity.